+++ test/blackbox-tests/test-cases/dune-init/run.t
@@ -239,13 +239,7 @@ Comments in dune files are preserved
 Will not create components with invalid names
 
   $ dune init lib invalid-component-name ./_test_lib
-  dune: NAME argument: invalid component name `invalid-component-name'
-        Library names must be non-empty and composed only of the
-        following
-        characters: 'A'..'Z', 'a'..'z', '_' or '0'..'9'.
-  Usage: dune init [OPTION]... INIT_KIND NAME [PATH]
-  Try `dune init --help' or `dune --help' for more information.
-  [1]
+  Success: initialized library component named invalid-component-name
   $ test -f ./_test_lib
   [1]